finito

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: amontalb <amontalb@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 15:31:43 by amontalb          #+#    #+#             */
-/*   Updated: 2022/11/09 09:45:58 by amontalb         ###   ########.fr       */
+/*   Updated: 2022/11/18 11:18:58 by amontalb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
